feat: Upgrade ElevenLabs default TTS model to `eleven_multilingual_v2`, remove deprecated models, and streamline avatar example project.
