+++ examples/avatar/index.html
@@ -16,16 +16,3 @@
   </body>
 </html>
 
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>elizaOS VRM Voice Demo</title>
-  </head>
-  <body>
-    <div id="root">Loading…</div>
-    <script type="module" src="/src/main.tsx"></script>
-  </body>
-</html>
-
